requirements.txt
certifi==2020.12.5
ciso8601==2.1.3
influxdb-client==1.15.0
paho-mqtt==1.5.1
pkg-resources==0.0.0
pylogix==0.7.8
python-dateutil==2.8.1
pytz==2021.1
Rx==3.1.1
six==1.15.0
urllib3==1.26.4
